+++ back/src/main/java/com/DatabaseInitializer.java
@@ -32,6 +32,8 @@ import com.unit.UnitRepository;
 import com.user.User;
 import com.user.UserRepository;
 
+import com.image.Image;
+import com.image.ImageRepository;
 import org.aspectj.weaver.ast.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -71,6 +73,7 @@ public class DatabaseInitializer {
   
         @Autowired
         private CourseRepository courseRepository;
+        
 
         @PostConstruct
         public void init() {
@@ -151,12 +154,12 @@ public class DatabaseInitializer {
                         "imperativos y OO\n");
 
                 slide03.addContent("\n" +
-                        "assert.card/1/1\n" +
+                        "insert.card/Definición/Array\n" +
                         "\n" +
-                        "assert.card/2/1\n");
+                        "insert.card/2/1\n");
 
                 slide04.addContent("\n" +
-                        "assert.slide/1/1/1\n");
+                        "insert.slide/Paradigmas de la programación/Introducción a la programación funcional/Programación funcional\n");
 
                 slideRepository.save(slide1);
                 slideRepository.save(slide01);