+++ back/src/main/java/com/slide/SlideRestController.java
@@ -1,5 +1,6 @@
 package com.slide;
 
+import java.util.List;
 import java.util.Optional;
 
 import com.GeneralRestController;
@@ -8,14 +9,7 @@ import com.card.Card;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.json.MappingJacksonValue;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/slides")
@@ -67,6 +61,17 @@ public class SlideRestController extends GeneralRestController {
         }
 
         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+    @GetMapping(value = "/search")
+    public ResponseEntity<List<Slide>> getSlideByName(@RequestParam String unitName, @RequestParam String lessonName, @RequestParam String slideName) {
+        System.out.println(unitName);
+        System.out.println(lessonName);
+        System.out.println(slideName);
+        List<Slide> slides = this.slideService.findByName(unitName, lessonName, slideName);
+        if (slides.size() == 0) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+        return new ResponseEntity<>(slides, HttpStatus.OK);
     }
 
 }