+++ front/src/app/app.module.ts
@@ -68,6 +68,9 @@ import {UnitModuleService} from './itinerary/module/unit-module.service';
 import {ModuleFormComponent} from './itinerary/module/moduleForm/module-form.component';
 import {UnitsBlocksToolComponent} from './itinerary/module/moduleEditor/units-blocks-tool.component';
 import {ModuleRenameComponent} from './itinerary/module/moduleRename/module-rename.component';
+import {ImageComponent} from './images/image.component';
+import {ImageService} from './images/image.service';
+import {LessonSlidesToolComponent} from './itinerary/lesson/lessonTools/lesson-slides-tool.component';
 import {QuestionTrackingComponent} from './question/questionTracking/questionTracking.component';
 
 @NgModule({
@@ -75,6 +78,7 @@ import {QuestionTrackingComponent} from './question/questionTracking/questionTra
     AppComponent,
     LessonFormComponent,
     LessonComponent,
+    LessonSlidesToolComponent,
     ModuleComponent,
     ModuleFormComponent,
     LoginComponent,
@@ -82,6 +86,7 @@ import {QuestionTrackingComponent} from './question/questionTracking/questionTra
     ViewComponent,
     UnitComponent,
     CardComponent,
+    ImageComponent,
     LessonEditorComponent,
     ModuleEditorComponent,
     UnitsCardsToolComponent,
@@ -156,7 +161,7 @@ import {QuestionTrackingComponent} from './question/questionTracking/questionTra
   ],
   providers: [LoginService, LessonService, ModuleService, CardService, SlideService, DefinitionQuestionService, ListQuestionService,
     QuestionService, ProgressService, TestQuestionService, MyCoursesService, CourseService, TabService, UnitService, NewCourseService,
-    MenuComponent, UnitLessonService, UnitModuleService,
+    ImageService, MenuComponent, UnitLessonService, UnitModuleService,
     { provide: HTTP_INTERCEPTORS, useClass: BasicAuthInterceptor, multi: true },
     { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
     { provide: LocationStrategy, useClass: HashLocationStrategy}],