+++ front/src/app/itinerary/lesson/lessonEditor/lesson-editor.component.ts
@@ -20,6 +20,10 @@ import {CourseService} from '../../../course/course.service';
 import {Course} from '../../../course/course.model';
 import {ModuleService} from '../../module/module.service';
 import {Module} from '../../module/module.model';
+import {ImageService} from '../../../images/image.service';
+import {ImageComponent} from '../../../images/image.component';
+import {CardService} from '../../../card/card.service';
+import {LessonSlidesToolComponent} from '../lessonTools/lesson-slides-tool.component';
 
 
 function convertToHTML(text) {
@@ -76,7 +80,9 @@ export class LessonEditorComponent implements OnInit {
               private moduleService: ModuleService,
               private bottomSheet: MatBottomSheet,
               private unitLessonService: UnitLessonService,
-              private tabService: TabService) {
+              private tabService: TabService,
+              private imageService: ImageService,
+              private cardService: CardService) {
     this.showSpinner = true;
   }
 
@@ -157,21 +163,69 @@ export class LessonEditorComponent implements OnInit {
     }
   }
 
-  async getEmbebedContent(contentId: number, contentId2: number, unitId: number, content: string, contentCounter: number, type: string) {
+  async getEmbebedContent(content1: any, content2: any, unit: any, content: string, contentCounter: number, type: string) {
     let contentEmbebed;
-    if (type === 'card') {
-      contentEmbebed = await this.unitService.getCard(contentId, unitId).toPromise();
-      this.extractedData.splice(contentCounter, 1, contentEmbebed.content);
-    } else if (type === 'slide') {
-      this.subSlide = true;
-      contentEmbebed = await this.unitLessonService.getSlideFormLesson(contentId, contentId2, unitId).toPromise();
-      this.extractedData.splice(contentCounter, 1, contentEmbebed.content.split('=== ')[1]);
-    } else if (type === 'question') {
-      contentEmbebed = await this.definitionQuestionService.getDefinitionQuestion(contentId).toPromise();
-      const url = 'http://localhost:4200/#/units/' + unitId + '/itineraries/11/definitionQuestion/' + contentId;
-      this.extractedData.splice(contentCounter, 1, contentEmbebed.questionText + '\n\n- ' + url + '[Resolver^]');
+    if (+unit) {
+      if (type === 'card') {
+        contentEmbebed = await this.unitService.getCard(+content1, +unit).toPromise().catch((error) => console.log(error));
+        if (typeof contentEmbebed !== 'undefined') {
+          this.extractedData.splice(contentCounter, 1, contentEmbebed.content);
+        } else {
+          this.extractedData.splice(contentCounter, 1, 'ERROR\n');
+        }
+      } else if (type === 'slide') {
+        this.subSlide = true;
+        contentEmbebed = await this.unitLessonService.getSlideFormLesson(+content1, +content2, +unit).toPromise().catch((error) => console.log(error));
+        if (typeof contentEmbebed !== 'undefined') {
+          this.extractedData.splice(contentCounter, 1, contentEmbebed.content.split('=== ')[1]);
+        } else {
+          this.extractedData.splice(contentCounter, 1, 'ERROR\n');
+        }
+      } else if (type === 'question') {
+        /* contentEmbebed = await this.definitionQuestionService.getDefinitionQuestion(content1).toPromise();
+        const url = 'http://localhost:4200/#/units/' + unit + '/itineraries/11/definitionQuestion/' + content1;
+        this.extractedData.splice(contentCounter, 1, contentEmbebed.questionText + '\n\n- ' + url + '[Resolver^]'); */
+      }
+    } else {
+      if (type === 'card') {
+        contentEmbebed = await this.cardService.getCardByName(unit, content1).toPromise().catch((error) => console.log(error));
+        if ((typeof contentEmbebed === 'undefined') || (contentEmbebed.length > 1)) {
+          this.extractedData.splice(contentCounter, 1, 'ERROR\n');
+        } else {
+          this.extractedData.splice(contentCounter, 1, contentEmbebed[0].content);
+        }
+      } else if (type === 'slide') {
+        this.subSlide = true;
+        contentEmbebed = await this.slideService.getSlideByName(unit, content2, content1).toPromise().catch((error) => console.log(error));
+        if ((typeof contentEmbebed === 'undefined') || (contentEmbebed.length > 1)) {
+          this.extractedData.splice(contentCounter, 1, 'ERROR\n');
+        } else {
+          this.extractedData.splice(contentCounter, 1, contentEmbebed[0].content.split('=== ')[1]);
+        }
+      }
     }
+
+    if (type === 'image') {
+      contentEmbebed = await this.imageService.getImage(content1).toPromise().catch((error) => console.log(error));
+      if (typeof contentEmbebed !== 'undefined') {
+        const image = this.convertImage(contentEmbebed.image);
+        const img = '++++\n' +
+          '<img class="img-lesson" src="' + image + '">\n' +
+          '++++\n' +
+          '\n';
+        this.extractedData.splice(contentCounter, 1, img);
+      } else {
+        this.extractedData.splice(contentCounter, 1, 'ERROR\n');
+      }
+    }
+
     this.addExtractedData(content);
+  }
+
+  convertImage(bytes: any) {
+    return 'data:image/png;base64,' + btoa(new Uint8Array(bytes).reduce((data, byte) =>
+      data + String.fromCharCode(byte),
+      ''));
   }
 
   contentCounterFunction(content: string) {
@@ -181,7 +235,7 @@ export class LessonEditorComponent implements OnInit {
     lines.forEach((line: string) => {
       let words: string[];
       words = line.split('.');
-      if (words[0] === 'assert') {
+      if (words[0] === 'insert') {
         this.contentCount = this.contentCount + 1;
       }
     });
@@ -202,20 +256,24 @@ export class LessonEditorComponent implements OnInit {
     lines.forEach((line: string) => {
       let words: string[];
       words = line.split('.');
-      if (words[0] === 'assert') {
+      if (words[0] === 'insert') {
         let parameters: string[];
         parameters = words[1].split('/');
         if (parameters[0] === 'card') {
           this.position.push(counter);
-          this.getEmbebedContent(Number(parameters[1]), null, Number(parameters[2]), content, contentCounter, 'card');
+          this.getEmbebedContent(parameters[1], null, parameters[2], content, contentCounter, 'card');
           contentCounter = contentCounter + 1;
         } else if (parameters[0] === 'slide') {
           this.position.push(counter);
-          this.getEmbebedContent(Number(parameters[1]), Number(parameters[2]), Number(parameters[3]), content, contentCounter, 'slide');
+          this.getEmbebedContent(parameters[1], parameters[2], parameters[3], content, contentCounter, 'slide');
           contentCounter = contentCounter + 1;
         } else if (parameters[0] === 'question') {
           this.position.push(counter);
           this.getEmbebedContent(Number(parameters[1]), null, Number(parameters[2]), content, contentCounter, 'question');
+          contentCounter = contentCounter + 1;
+        } else if (parameters[0] === 'image') {
+          this.position.push(counter);
+          this.getEmbebedContent(Number(parameters[1]), null, null, content, contentCounter, 'image');
           contentCounter = contentCounter + 1;
         }
       }
@@ -257,8 +315,9 @@ export class LessonEditorComponent implements OnInit {
     if (slidesContent[0].split(' ')[0] === '==') {
       this.lesson.name = '';
       for (let i = 1; i < slidesContent[0].split(' ').length; i ++) {
-        this.lesson.name = this.lesson.name + slidesContent[0].split(' ')[i].split('\n')[0] + ' ';
+        this.lesson.name = this.lesson.name + slidesContent[0].split(' ')[i] + ' ';
       }
+      this.lesson.name = this.lesson.name.split('\n')[0];
     }
     this.contentToSlides(slidesContent);
   }
@@ -299,8 +358,16 @@ export class LessonEditorComponent implements OnInit {
     });
   }
 
-  openBottomSheet(): void {
+  openCardsBottomSheet(): void {
     this.bottomSheet.open(UnitsCardsToolComponent);
+  }
+
+  openImageBottomSheet(): void {
+    this.bottomSheet.open(ImageComponent);
+  }
+
+  openSlidesBottomSheet(): void {
+    this.bottomSheet.open(LessonSlidesToolComponent);
   }
 
   nextSlide() {